move is_overridden to model_helpers
