+++ energizer/loops/pool_epoch_loop.py
@@ -1,15 +1,13 @@
-from functools import lru_cache, partial
-from typing import Any, Callable, Optional, Type
-from unittest.mock import Mock
+from functools import lru_cache
+from typing import Any, Callable, Optional
 
-import pytorch_lightning as pl
 import torch
 from pytorch_lightning.loops.epoch.evaluation_epoch_loop import EvaluationEpochLoop
 from pytorch_lightning.utilities.types import STEP_OUTPUT
 from torch import Tensor
 from torchmetrics import Metric
 
-from energizer.learners.base import Learner
+from energizer.utilities.model_helpers import is_overridden
 
 
 class AccumulateTopK(Metric):
@@ -130,41 +128,3 @@ class PoolEvaluationEpochLoop(EvaluationEpochLoop):
         """Whether the batch outputs should be stored for later usage."""
         model = self.trainer.lightning_module
         return is_overridden("pool_epoch_end", model)
-
-
-def is_overridden(method_name: str, instance: Optional[object] = None, parent: Optional[Type[object]] = None) -> bool:
-    if instance is None:
-        # if `self.lightning_module` was passed as instance, it can be `None`
-        return False
-
-    if parent is None:
-        if isinstance(instance, Learner):
-            parent = Learner
-        elif isinstance(instance, pl.LightningDataModule):
-            parent = pl.LightningDataModule
-        elif isinstance(instance, pl.Callback):
-            parent = pl.Callback
-        if parent is None:
-            raise ValueError("Expected a parent")
-
-    instance_attr = getattr(instance, method_name, None)
-    if instance_attr is None:
-        return False
-    # `functools.wraps()` support
-    if hasattr(instance_attr, "__wrapped__"):
-        instance_attr = instance_attr.__wrapped__
-    # `Mock(wraps=...)` support
-    if isinstance(instance_attr, Mock):
-        # access the wrapped function
-        instance_attr = instance_attr._mock_wraps
-    # `partial` support
-    elif isinstance(instance_attr, partial):
-        instance_attr = instance_attr.func
-    if instance_attr is None:
-        return False
-
-    parent_attr = getattr(parent, method_name, None)
-    if parent_attr is None:
-        raise ValueError("The parent should define the method")
-
-    return instance_attr.__code__ != parent_attr.__code__
